feat: add types to service classes
The very basic implementation so far. There still unresolved issues:
cyclic dependencies, no type on the arguments coming from HTTP request.
